sync

+++ scripts/osfstorage/files_audit.py
@@ -10,18 +10,21 @@ Note: Must have par2 installed to run
 
 from __future__ import division
 
+import gc
 import os
 import math
+import hashlib
 import logging
 
 import pyrax
-import progressbar
-from modularodm import Q
+
 from boto.glacier.layer2 import Layer2
 from pyrax.exceptions import NoSuchObject
 
+from framework.celery_tasks import app as celery_app
+
 from website.app import init_app
-from website.addons.osfstorage import model
+from osf.models import FileVersion
 
 from scripts import utils as scripts_utils
 from scripts.osfstorage import utils as storage_utils
@@ -31,34 +34,49 @@ from scripts.osfstorage import settings as storage_settings
 container_primary = None
 container_parity = None
 vault = None
+audit_temp_path = None
+
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
 logging.getLogger('boto').setLevel(logging.CRITICAL)
 
 
+def delete_temp_file(version):
+    path = os.path.join(audit_temp_path, version.location['object'])
+    try:
+        os.remove(path)
+    except OSError:
+        pass
+
+
 def download_from_cloudfiles(version):
-    path = os.path.join(storage_settings.AUDIT_TEMP_PATH, version.location['object'])
+    path = os.path.join(audit_temp_path, version.location['object'])
     if os.path.exists(path):
-        return path
+        # we cannot assume the file is valid and not from a previous failure.
+        delete_temp_file(version)
     try:
         obj = container_primary.get_object(version.location['object'])
-        obj.download(storage_settings.AUDIT_TEMP_PATH)
+        with open(path, 'wb') as fp:
+            hasher = hashlib.sha256()
+            fetcher = obj.fetch(chunk_size=262144000)  # 256mb chunks
+            while True:
+                try:
+                    chunk = next(fetcher)
+                except StopIteration:
+                    break
+                hasher.update(chunk)
+                fp.write(chunk)
+        if hasher.hexdigest() != version.metadata['sha256']:
+            raise Exception('SHA256 mismatch, cannot continue')
+        return path
     except NoSuchObject as err:
         logger.error('*** FILE NOT FOUND ***')
         logger.error('Exception:')
         logger.exception(err)
         logger.error('Version info:')
         logger.error(version.to_storage())
-    return path
-
-
-def delete_temp_file(version):
-    path = os.path.join(storage_settings.AUDIT_TEMP_PATH, version.location['object'])
-    try:
-        os.remove(path)
-    except OSError:
-        pass
+        return None
 
 
 def ensure_glacier(version, dry_run):
@@ -67,11 +85,11 @@ def ensure_glacier(version, dry_run):
     logger.warn('Glacier archive for version {0} not found'.format(version._id))
     if dry_run:
         return
-    download_from_cloudfiles(version)
-    file_path = os.path.join(storage_settings.AUDIT_TEMP_PATH, version.location['object'])
-    glacier_id = vault.upload_archive(file_path, description=version.location['object'])
-    version.metadata['archive'] = glacier_id
-    version.save()
+    file_path = download_from_cloudfiles(version)
+    if file_path:
+        glacier_id = vault.upload_archive(file_path, description=version.location['object'])
+        version.metadata['archive'] = glacier_id
+        version.save()
 
 
 def check_parity_files(version):
@@ -87,46 +105,55 @@ def ensure_parity(version, dry_run):
     if dry_run:
         return
     file_path = download_from_cloudfiles(version)
-    parity_paths = storage_utils.create_parity_files(file_path)
-    for parity_path in parity_paths:
-        container_parity.create(parity_path)
-        os.remove(parity_path)
-    if not check_parity_files(version):
-        logger.error('Parity files for version {0} not found after update'.format(version._id))
+    if file_path:
+        parity_paths = storage_utils.create_parity_files(file_path)
+        for parity_path in parity_paths:
+            container_parity.create(parity_path)
+            os.remove(parity_path)
+        if not check_parity_files(version):
+            logger.error('Parity files for version {0} not found after update'.format(version._id))
 
 
 def ensure_backups(version, dry_run):
-    if version.size == 0:
-        return
     ensure_glacier(version, dry_run)
     ensure_parity(version, dry_run)
     delete_temp_file(version)
 
 
-def get_targets():
-    return model.OsfStorageFileVersion.find(
-        Q('status', 'ne', 'cached') &
-        Q('location.object', 'exists', True)
-    )
+def glacier_targets():
+    return FileVersion.objects.filter(location__has_key='object', metadata__archive__isnull=True)
 
 
-def main(nworkers, worker_id, dry_run):
-    targets = get_targets()
-    progress_bar = progressbar.ProgressBar(maxval=math.ceil(targets.count() / nworkers)).start()
+def parity_targets():
+    # TODO: Add metadata.parity information from wb so we do not need to check remote services
+    return FileVersion.objects.filter(location__has_key='object')
+        # & metadata__parity__isnull=True
+
+
+def audit(targets, num_of_workers, worker_id, dry_run):
+    maxval = math.ceil(targets.count() / num_of_workers)
+    target_iterator = targets.iterator()
     idx = 0
-    for version in targets:
-        if hash(version._id) % nworkers == worker_id:
+    last_progress = -1
+    for version in target_iterator:
+        if hash(version._id) % num_of_workers == worker_id:
+            if version.size == 0:
+                continue
             ensure_backups(version, dry_run)
             idx += 1
-            progress_bar.update(idx)
-    progress_bar.finish()
+            progress = int(idx / maxval * 100)
+            if last_progress < 100 and last_progress < progress:
+                logger.info(str(progress) + '%')
+                last_progress = progress
+                gc.collect()
 
 
-if __name__ == '__main__':
-    import sys
-    nworkers = int(sys.argv[1])
-    worker_id = int(sys.argv[2])
-    dry_run = 'dry' in sys.argv
+@celery_app.task(name='scripts.osfstorage.files_audit')
+def main(num_of_workers=0, worker_id=0, glacier=True, parity=True, dry_run=True):
+    global container_primary
+    global container_parity
+    global vault
+    global audit_temp_path
 
     # Set up storage backends
     init_app(set_backends=True, routes=False)
@@ -152,8 +179,35 @@ if __name__ == '__main__':
         # Log to file
         if not dry_run:
             scripts_utils.add_file_logger(logger, __file__, suffix=worker_id)
-            main(nworkers, worker_id, dry_run=dry_run)
+
+        audit_temp_path = os.path.join(storage_settings.AUDIT_TEMP_PATH, str(worker_id))
+        if not dry_run:
+            try:
+                os.makedirs(audit_temp_path)
+            except OSError:
+                pass
+
+        if glacier:
+            logger.info('glacier audit start')
+            audit(glacier_targets(), num_of_workers, worker_id, dry_run)
+            logger.info('glacier audit complete')
+
+        if parity:
+            logger.info('parity audit start')
+            audit(parity_targets(), num_of_workers, worker_id, dry_run)
+            logger.info('parity audit complete')
+
     except Exception as err:
         logger.error('=== Unexpected Error ===')
         logger.exception(err)
         raise err
+
+
+if __name__ == '__main__':
+    import sys
+    arg_num_of_workers = int(sys.argv[1])
+    arg_worker_id = int(sys.argv[2])
+    arg_glacier = 'glacier' in sys.argv
+    arg_parity = 'parity' in sys.argv
+    arg_dry_run = 'dry' in sys.argv
+    main(num_of_workers=arg_num_of_workers, worker_id=arg_worker_id, glacier=arg_glacier, parity=arg_parity, dry_run=arg_dry_run)
